feat(poa/ante): unify setup poa decorator
