+++ x/poa/ante/poa_test.go
@@ -2,12 +2,28 @@ package ante
 
 import (
 	"testing"
+	"time"
 
+	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
+	sdktestutil "github.com/cosmos/cosmos-sdk/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 	"github.com/xrplevm/node/v3/x/poa/testutil"
+	"github.com/xrplevm/node/v3/x/poa/types"
 )
+
+func setupPoaDecorator(t *testing.T) (
+	*PoaDecorator,
+	sdk.Context,
+) {
+	key := sdk.NewKVStoreKey(types.StoreKey)
+	tsKey := sdk.NewTransientStoreKey("transient_test")
+	testCtx := sdktestutil.DefaultContextWithDB(t, key, tsKey)
+	ctx := testCtx.Ctx.WithBlockHeader(tmproto.Header{Time: time.Now()})
+
+	return &PoaDecorator{}, ctx
+}
 
 func TestPoaDecorator_AnteHandle(t *testing.T) {
 	pd, ctx := setupPoaDecorator(t)
